+++ lib/entity/transaction/extensions/default/recurring.g.dart
@@ -8,6 +8,9 @@ part of 'recurring.dart';
 
 Recurring _$RecurringFromJson(Map<String, dynamic> json) => Recurring(
   uuid: json['uuid'] as String,
+  initialTransactionDate: DateTime.parse(
+    json['initialTransactionDate'] as String,
+  ),
   relatedTransactionUuid: json['relatedTransactionUuid'] as String?,
   locked: json['locked'] as bool? ?? false,
 );
@@ -15,6 +18,7 @@ Recurring _$RecurringFromJson(Map<String, dynamic> json) => Recurring(
 Map<String, dynamic> _$RecurringToJson(Recurring instance) => <String, dynamic>{
   'uuid': instance.uuid,
   'relatedTransactionUuid': instance.relatedTransactionUuid,
+  'initialTransactionDate': instance.initialTransactionDate.toIso8601String(),
   'locked': instance.locked,
   'key': instance.key,
 };