+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [OPTIMISTIC] Now you can setup credit line accounts.
 * Added list of contributors, closes [#448](https://github.com/flow-mn/flow/issues/448)
 * Date and time are now in Arabic when the language is set to Arabic, thanks to [Ultrate](https://github.com/Ultrate)
+* Now you can do recurring transactions, closes [#59](https://github.com/flow-mn/flow/issues/59)
 
 ### Changes & Improvements